ADS_virtual_lab-main exp 4/api/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pandas==2.3.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.4
python-dotenv==1.0.1
scipy==1.14.1
